refactor(sqlite): Extract and compile generator
